backend/requirements.txt
fastapi>=0.115.0,<1.0.0
uvicorn[standard]>=0.30.0,<1.0.0
langgraph>=0.2.40
langchain-core>=0.3.0
langchain>=0.3.0
langchain-groq>=0.2.0
langchain-community>=0.3.0
tavily-python>=0.5.0
pydantic-settings>=2.6.0
python-dotenv>=1.0.1
httpx>=0.27.0
pytest>=8.3.0
